enhancement(constructors): Only insert provided attributes
Previous method left unbound parameters in the query if not all attributes
were provided, which are interpreted by SQLite as `NULL` values.

By only crafting the `INSERT` statement with the provided attributes, all
variables of the prepared statement are bound, putting the onus on the DB
to do the needful (e.g. fill with defaults, raise constraint error).